--- app/src/main/java/hu/tb/minichefy/domain/model/storage/CalculationFood.kt
@@ -1,18 +1,0 @@
-package hu.tb.minichefy.domain.model.storage
-
-import hu.tb.minichefy.domain.model.storage.UnitOfMeasurement.*
-
-data class CalculationFood(
-    val quantity: Float,
-    val unitOfMeasurement: UnitOfMeasurement
-) {
-    fun convertToSmallestUnit() =
-        when (unitOfMeasurement) {
-            KG -> quantity * 1000
-            DKG -> quantity * 10
-            G -> quantity
-            L -> quantity * 10
-            DL -> quantity
-            PIECE -> quantity
-        }
-}
