+++ app/src/main/java/hu/tb/minichefy/domain/use_case/Validators.kt
@@ -1,7 +1,6 @@
 package hu.tb.minichefy.domain.use_case
 
 class ValidateQuantityNumber {
-
     operator fun invoke(quantityNumber: Int): ValidationResult =
         if (quantityNumber < 1) {
             ValidationResult.ERROR
