+++ app/src/main/java/hu/tb/minichefy/presentation/preview/ProductPreviewParameterProvider.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import hu.tb.minichefy.domain.model.storage.Food
 import hu.tb.minichefy.domain.model.storage.FoodTag
 import hu.tb.minichefy.domain.model.storage.UnitOfMeasurement
-import hu.tb.minichefy.presentation.screens.components.icons.IconManager
+import hu.tb.minichefy.presentation.screens.manager.icons.IconManager
 
 class ProductPreviewParameterProvider : PreviewParameterProvider<List<Food>> {
