+++ app/src/main/java/hu/tb/minichefy/presentation/screens/components/IconSelectorSheet.kt
@@ -31,9 +31,9 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import hu.tb.minichefy.presentation.screens.components.icons.IconManager
-import hu.tb.minichefy.presentation.screens.components.icons.IconResource
-import hu.tb.minichefy.presentation.screens.components.icons.MealIcon
+import hu.tb.minichefy.presentation.screens.manager.icons.IconManager
+import hu.tb.minichefy.presentation.screens.manager.icons.IconResource
+import hu.tb.minichefy.presentation.screens.manager.icons.MealIcon
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
