--- app/src/main/java/hu/tb/minichefy/presentation/screens/components/icons/IconResource.kt
@@ -1,5 +1,0 @@
-package hu.tb.minichefy.presentation.screens.components.icons
-
-interface IconResource {
-    val resource: Int
-}
