+++ app/src/main/java/hu/tb/minichefy/presentation/screens/manager/icons/IconResource.kt
@@ -1,0 +1,5 @@
+package hu.tb.minichefy.presentation.screens.manager.icons
+
+interface IconResource {
+    val resource: Int
+}
