+++ app/src/main/java/hu/tb/minichefy/presentation/screens/storage/storage_list/componenets/EditStorageItem.kt
@@ -44,7 +44,7 @@ import hu.tb.minichefy.R
 import hu.tb.minichefy.domain.model.storage.Food
 import hu.tb.minichefy.domain.model.storage.FoodTag
 import hu.tb.minichefy.presentation.preview.ProductPreviewParameterProvider
-import hu.tb.minichefy.presentation.screens.components.icons.iconVectorResource
+import hu.tb.minichefy.presentation.screens.manager.icons.iconVectorResource
 import hu.tb.minichefy.presentation.ui.components.clickableWithoutRipple
 import hu.tb.minichefy.presentation.ui.theme.Green500
 import hu.tb.minichefy.presentation.ui.theme.MEDIUM_SPACE_BETWEEN_ELEMENTS
